refactor: App에서 useCart 로직 분리
