+++ src/basic/hooks/useCart.ts
@@ -1,0 +1,37 @@
+import { useState, useEffect } from "react";
+import { CartItem } from "../../types";
+
+export const useCart = () => {
+  const [cart, setCart] = useState<CartItem[]>(() => {
+    const saved = localStorage.getItem("cart");
+    if (saved) {
+      try {
+        return JSON.parse(saved);
+      } catch {
+        return [];
+      }
+    }
+    return [];
+  });
+
+  const [totalItemCount, setTotalItemCount] = useState(0);
+
+  useEffect(() => {
+    if (cart.length > 0) {
+      localStorage.setItem("cart", JSON.stringify(cart));
+    } else {
+      localStorage.removeItem("cart");
+    }
+  }, [cart]);
+
+  useEffect(() => {
+    const count = cart.reduce((sum, item) => sum + item.quantity, 0);
+    setTotalItemCount(count);
+  }, [cart]);
+
+  return {
+    cart,
+    setCart,
+    totalItemCount,
+  };
+};
